changed
GetCallbackUrl has been moved to a separate static class, and used where necessary.
