+++ ExpenseTracker.Api/ExpenseTracker.Application/Services/CurrentUserService.cs
@@ -31,6 +31,7 @@ public sealed class CurrentUserService : ICurrentUserService
 
     public string GetUserName()
     {
+        return "";
         var user = _httpContextAccessor.HttpContext?.User
             ?? throw new InvalidOperationException($"Unable to get user info from HttpContext.");
